requirements.txt
Django==1.11
django-filter==1.0.4
djangorestframework==3.6.3
gunicorn==19.7.1
psycopg2==2.7.1
django-storages==1.6.3
boto==2.48.0
Pillow==4.2.1

kong_oauth==0.1

ipdb
ipython

coverage
codacy-coverage
Faker
